* EOAccess/EOEntity.m           remove unused variable         * EOControl/EOEditingContext.m           use the object, instead of the EOGlobalID to get the entityName         * EOControl/EOGenericRecord.m         * EOInterface/EOColumnAssociation.[h|m]         * EOInterface/EOTableViewAssociation.m           some NS(U)Integer conversions
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gdl2/trunk@36677 72102866-910b-0410-8b05-ffd578937521

+++ EOInterface/EOColumnAssociation.h
@@ -91,20 +91,20 @@
 - (void)tableView: (NSTableView *)tableView
    setObjectValue: (id)object
    forTableColumn: (NSTableColumn *)tableColumn
-	      row: (int)row;
+	      row: (NSInteger)row;
 
 - (id)tableView: (NSTableView *)tableView
 objectValueForTableColumn: (NSTableColumn *)tableColumn
-	    row: (int)row;
+	    row: (NSInteger)row;
 
 - (BOOL)tableView: (NSTableView *)tableView
 shouldEditTableColumn: (NSTableColumn *)tableColumn
-	      row: (int)row;
+	      row: (NSInteger)row;
 
 - (void)tableView: (NSTableView *)tableView
   willDisplayCell: (id)cell
    forTableColumn: (NSTableColumn *)tableColumn
-	      row: (int)row;
+	      row: (NSInteger)row;
 
 /* EOColumnViewAssociation control delegate.  */
 - (BOOL)control: (NSControl *)control
@@ -169,21 +169,21 @@ textShouldBeginEditing: (NSText *)fieldEditor;
 - (void)tableView: (NSTableView *)tableView
    setObjectValue: (id)object
    forTableColumn: (NSTableColumn *)tableColumn
-	      row: (int)row;
+	      row: (NSInteger)row;
 
 - (id)tableView: (NSTableView *)tableView
 objectValueForTableColumn: (NSTableColumn *)tableColumn
-	    row: (int)row;
+	    row: (NSInteger)row;
 
 /* Delegate methods for table view.  */
 - (BOOL)tableView: (NSTableView *)tableView
 shouldEditTableColumn: (NSTableColumn *)tableColumn
-	      row: (int)row;
+	      row: (NSInteger)row;
 
 - (void)tableView: (NSTableView *)tableView
   willDisplayCell: (id)cell
    forTableColumn: (NSTableColumn *)tableColumn
-	      row: (int)row;
+	      row: (NSInteger)row;
 
 /* Notification methods for table view.  */
 - (void)tableViewSelectionDidChange: (NSNotification *)notification;
